Testing Sprite Movement, Collisions, & Camera Zoom
Another day.
I continued following along with the Youtube tutorial produced by
C4Cheats Games, and implemented some basic sprite movement. Additionally
a hidden rectangle was added into the space to test collisions.

The channel playlist can be found here:

https://www.youtube.com/watch?v=WWSF96vH8vs&index=4&list=PLvnEciYMRSk-l5oKsIsbrBbP2ZJFSZLns

A zooming camera is something that I've been wanting to implement, and I
got really curious about trying it out while watching C4Cheats Games'
videos. I Googled-around and was pretty quickly informed that I needed
to create a custom class which implemented the libGDX InputProcessor
interface. I then needed to "tell" libGDX that an InputProcessor existed
and where to find it. I accomplished this by quickly making a "testCam"
class which extended the OrthographicCamera class and implemented this
InputProcessor interface. Below is the wiki page explaining how to
manipulate the zoom of an OrthographicCamera, and thankfully it was
written in very clear language with some convenient code examples. I
combined this with the "scolled" method given by InputProcessor, and was
pleased to see zooming-in/out work almost instantly.

https://github.com/libgdx/libgdx/wiki/Orthographic-camera