Incremental Improvements
On-screen text now displays FPS, world-coordinates on which the camera
is centered, and the world-coordinates of the mouse pointer.

When a Mill entity is clicked the camera is centered over its position,
and the camera zoom is adjusted appropriately. The output terminal
confirms the detection of the unique sprite.

A grid now displays on-screen which is described by the LineGrid class.
It can be arbitrarily placed in the game world, given values for width
and height in terms of grid spaces, and can also be given a
pixel-size-per-space. The colors of the verticals and horizontals can be
adjusted by the user if desired. The grid is comprised of lines drawn by
the ShapeRenderer class which means that the fewest number of objects
necessary (continuous lines) are being used to create the grid.
Rectangles enclosing all four sides of the grid are also present, but
can be turned-off by setting a boolean value and calling the LineGrid's
draw() method once more. The rectangles are both a Rectangle object as
well as filled boxes drawn by the LineGrid's ShapeRenderer. The purpose
of these boxes is for any sort of collision detection compatible with
overlapping rectangles.